kernel: add 'struct proc' and basic context-switching support
Added save() and resume(), kernel analogs of the setjmp()/longjmp()
which are used for context-switching. Added 'struct proc' as a
container for the context required for these primitives.

We now craft proc0 on startup and resume() it to get the BSP on a
properly allocated and mapped kernel stack.

Added floating-point initialization to the CPU startup sequence.

Some "stray" changes appear here, which are just whitespace
adjustments (from TAB to space expansions) for style.

Signed-off-by: Charles E. Youse <charles@gnuless.org>