kernel: page.c: page accounting in pmap[]
When allocating a physical page, update the pmap[] to indicate what
the page is used for, and allow the caller to specify a type-specific
value to store in the pmap with that page. e.g., if it's a slab page,
mark it PMAP_SLAB and store a pointer to the slab in the pmap[].

This is both for accounting (how many pages are used for what?) and in
some cases for performace (we can quickly determine the virtual address
of a physical page in a process's address space).

Signed-off-by: Charles E. Youse <charles@gnuless.org>