kernel: slab.c: use static slab initialization
The values for a slab can be computed at build time, so do that
instead of requiring a call to an initializer at runtime.

Also, fix the arrangement of objects in a slab for efficiency:
previously, the slab_page header took up the first object's place
regardless of that object's size. For large objects like 'struct
proc' this wastes a lot of space. Instead, only reserve SLAB_MIN
bytes for the header and start placing objects immediately after.

Signed-off-by: Charles E. Youse <charles@gnuless.org>